pull MarkovNode and MarkovEdge into new modules
This commit just reorganizes the MarkovNode and MarkovEdge types into
their own submodules. This is another pre-factor for changing how the
gadgets are organized.

Test plan: Just a reorg, look for green CI.
